// frontend/src/api/openapi/models/LeagueFixtureResults.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { GameweekFixtureResult } from './GameweekFixtureResult';
export type LeagueFixtureResults = {
    gameweekFixtures: Array<GameweekFixtureResult>;
    leagueId: number;
};