// frontend/src/api/openapi/models/GameweekFixtureResult.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { PublicUser } from './PublicUser';
export type GameweekFixtureResult = {
    user1: PublicUser;
    user1Score: number;
    user2: PublicUser;
    user2Score: number;
    gameweekId: number;
    gameweekNumber: number;
    oddPlayer: boolean;
};